playing style recognition added and working with basic functionality

+++ smart_multi_effect/tobinary.h
@@ -7,8 +7,8 @@
 template<typename T>
 size_t VectorToBinary(std::ofstream& fout, std::vector<T> &vec)
 {
-    size_t vecSize = vec.size();
-    fout.write((char*)&vecSize, sizeof (size_t));
+    uint64_t vecSize = vec.size();
+    fout.write((char*)&vecSize, sizeof (uint64_t));
     fout.write((char*)&vec[0], vecSize * sizeof (T));
 
     return vecSize;
@@ -17,8 +17,8 @@ size_t VectorToBinary(std::ofstream& fout, std::vector<T> &vec)
 template<typename T>
 void BinaryToVector(std::ifstream &fin, std::vector<T> &vec)
 {
-    size_t vecSize = 0;
-    fin.read((char*)&vecSize, sizeof (size_t));
+    uint64_t vecSize = 0;
+    fin.read((char*)&vecSize, sizeof (uint64_t));
     vec.resize(vecSize);
     fin.read((char*)&vec[0], vecSize * sizeof (T));
 }
